% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{c.polynom}
\alias{c.polynom}
\alias{c.polylist}
\title{Concatenation of polynomial objects into lists}
\usage{
\method{c}{polynom}(..., recursive = FALSE)

\method{c}{polylist}(..., recursive = FALSE)
}
\arguments{
\item{...}{Polynomial or polylist objects}

\item{recursive}{Logical, should the concatenation flatten all component lists?}
}
\value{
A polylist object with all argumets included
}
\description{
Concatenation of polynomial objects into lists
}
