% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{change.origin}
\alias{change.origin}
\alias{change.origin.default}
\alias{change.origin.polynom}
\alias{change.origin.polylist}
\title{Chenge origin of a polynomial}
\usage{
change.origin(p, o, ...)

\method{change.origin}{default}(p, o, ...)

\method{change.origin}{polynom}(p, o, ...)

\method{change.origin}{polylist}(p, o, ...)
}
\arguments{
\item{p}{A polynom or polylist object}

\item{o}{A single numeric quantity specifying the new x-origin}

\item{...}{currently not used}
}
\value{
A polynom or polylist object with x measured from the new origin
}
\description{
Given a polynomial P(x) and a new origin \code{o}, find
the polynomial Q(x) = P(x + o).  I.e. Q(0) = P(o)
}
