% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{change_origin}
\alias{change_origin}
\alias{change_origin.default}
\alias{change_origin.polynom}
\alias{change_origin.polylist}
\title{Change origin of a polynomial}
\usage{
change_origin(p, o, ...)

\method{change_origin}{default}(p, o, ...)

\method{change_origin}{polynom}(p, o, ...)

\method{change_origin}{polylist}(p, o, ...)
}
\arguments{
\item{p}{A polynom or polylist object}

\item{o}{A single numeric quantity specifying the new x-origin}

\item{...}{currently not used}
}
\value{
A polynom or polylist object with x measured from the new origin
}
\description{
Given a polynomial P(x) and a new origin \code{o}, find
the polynomial Q(x) = P(x + o).  I.e. Q(0) = P(o)
}
