% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{coef.polynom}
\alias{coef.polynom}
\alias{coef.polylist}
\title{Polynomial coefficients}
\usage{
\method{coef}{polynom}(object, ...)

\method{coef}{polylist}(object, ...)
}
\arguments{
\item{object}{A polynomial object or list thereof}

\item{...}{Ignored}
}
\value{
A numeric vector of coefficients
}
\description{
Extract polynomial coefficients
}
\examples{
p <- polynomial(1:3)*polynomial(5:1)
coef(p)
}
