% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{defunct}
\alias{defunct}
\alias{as.polynom}
\alias{is.polynom}
\alias{is.polylist}
\alias{as.polylist}
\alias{change.origin}
\alias{poly.calc}
\alias{poly.from.zeros}
\alias{poly.from.values}
\title{Defunct functions}
\usage{
as.polynom(a)

is.polynom(a)

is.polylist(x)

as.polylist(x)

change.origin(p, o, ...)

poly.calc(...)

poly.from.zeros(...)

poly.from.values(...)
}
\arguments{
\item{a, p, x}{A numeric vector or polynom object}

\item{o}{A numeric vector of length one.}

\item{...}{Additional arguments of appropriate class}
}
\value{
An error is triggered and appropriate error message is returned
}
\description{
These functions have been removed from the package
to allow for systematic nomenclature.  Appropriate
drop-in replacements are given if the function is
called
}
