% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{deriv.polynom}
\alias{deriv.polynom}
\alias{integral}
\alias{integral.default}
\alias{integral.polynom}
\alias{deriv.polylist}
\alias{integral.polylist}
\title{Polynomial Calculus}
\usage{
\method{deriv}{polynom}(expr, ...)

integral(expr, ...)

\method{integral}{default}(expr, ...)

\method{integral}{polynom}(expr, limits = NULL, ...)

\method{deriv}{polylist}(expr, ...)

\method{integral}{polylist}(expr, ...)
}
\arguments{
\item{expr}{A polynomial object, or list thereof}

\item{...}{Unused as yet}

\item{limits}{Real limits of a definite integral}
}
\value{
A coeffieient vector, or list thereof
}
\description{
Find the derivative or indefinite integral of a polynomial object, or list thereof.
}
\examples{
p <- poly_from_roots(-2:3)
p
deriv(p)
integral(p)
}
