% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{LCM}
\alias{LCM}
\alias{lowest_common_multiple}
\alias{LCM.polynom}
\alias{LCM.polylist}
\title{Lowest Common Multiple}
\usage{
LCM(...)

lowest_common_multiple(...)

\method{LCM}{polynom}(...)

\method{LCM}{polylist}(...)
}
\arguments{
\item{...}{A list of polynomials or polylist objects}
}
\value{
A polynomial giving the lowest common multiple
}
\description{
For a list of polynomials, find the lowest degree monic
polynomial into which each divides exactly
}
\examples{
p <- poly_calc(0:5)
r <- poly_calc(1:6)
greatest_common_divisor(p, r)
solve(greatest_common_divisor(p, r))
lowest_common_multiple(p, r)
solve(lowest_common_multiple(p, r))
}
