\name{PolynomF-internal}
\title{Internal PolynomF objects}
\alias{.accumulate}
\alias{.degree}
\alias{.GCD2}
\alias{.LCM2}
\alias{.monic}
\alias{.poly.mult}
\alias{.poly.quo}
\alias{.poly.rem}
\alias{.polylist_from_list}
\description{Internal PolynomF objects.}
\details{These are not to be called by the user.}
\keyword{internal}