% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{predict.polynom}
\alias{predict.polynom}
\alias{predict.polylist}
\title{Evaluate a polynomial}
\usage{
\method{predict}{polynom}(object, newdata, ...)

\method{predict}{polylist}(object, newdata, ...)
}
\arguments{
\item{object}{A polynomial or polylist object}

\item{newdata}{A target object at which to evaluate.}

\item{...}{Not used}
}
\value{
If \code{newdata} is a numeric vector, a numeric vector of
        results.  If \code{newdata} is a polynomial, then the composition
        is returned as a polynomial, or polylist object.
}
\description{
Evaluate a polynomial, or polylist object components.
}
