% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{print.polylist}
\alias{print.polylist}
\title{Print method for polynomial objects}
\usage{
\method{print}{polylist}(x, ...)
}
\arguments{
\item{x}{A polynomial object or list thereof}

\item{...}{Additional arguments passed on to methods}
}
\value{
The original object, invisibly.
}
\description{
Print method for polynomial objects
}
