% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{print.polynom}
\alias{print.polynom}
\title{Print method for polynomial objects}
\usage{
\method{print}{polynom}(x, variable = "x", digits = getOption("digits"), decreasing = FALSE, ...)
}
\arguments{
\item{x}{A polynomial object}

\item{variable}{Character string: what variable name should be given?}

\item{digits}{Integer: how many decimal degits to use?}

\item{decreasing}{Logical: in descending powers, or ascending?}

\item{...}{Additional arguments}
}
\value{
The original object \code{x}, invisibly
}
\description{
Standard method for printing polynomial objects
}
