% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{rep.polylist}
\alias{rep.polylist}
\alias{rep.polynom}
\title{Component repition}
\usage{
\method{rep}{polylist}(x, times, ...)

\method{rep}{polynom}(x, times, ...)
}
\arguments{
\item{x}{A single polynom or polylist object}

\item{times, ...}{As for the base package function \code{rep}.}
}
\value{
The resulting polylist object.
}
\description{
Repeat components of a polylist object
}
