% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{solve.polynom}
\alias{solve.polynom}
\alias{solve.polylist}
\title{Find Polynomial Zeros}
\usage{
\method{solve}{polynom}(a, b, ...)

\method{solve}{polylist}(a, b, ...)
}
\arguments{
\item{a, b}{Polynomials for the LHS and RHS respectively}

\item{...}{Currently unused}
}
\value{
A vector of roots, usually complex
}
\description{
Solve polynomial equations, a(x) = b(x), or alternatively
find the zeros of the polynomial a(x) - b(x)
}
\examples{
p <- poly_calc(0:5)
solve(p)
solve(p, 1)
}
