% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{[.polylist}
\alias{[.polylist}
\title{Extract components of a list of polynomials}
\usage{
\method{[}{polylist}(x, i)
}
\arguments{
\item{x}{A polylist object}

\item{i}{An index vector of any crongruent form}
}
\value{
A polylist object of the components
}
\description{
Extract components of a list of polynomials
}
