% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{summary.polynom}
\alias{summary.polynom}
\alias{summary.polylist}
\alias{print.summary.polynom}
\title{Polynomial summary}
\usage{
\method{summary}{polynom}(object, ...)

\method{summary}{polylist}(object, ...)

\method{print}{summary.polynom}(x, ...)
}
\arguments{
\item{object, x}{A polynomial or polylist object}

\item{...}{Currently unused}
}
\value{
A list giving the zeros, stationary points and points of inflexion of the polynomial(s)
}
\description{
Provide a succinct summary of the critical points of a polynomial, or list thereof
}
\examples{
p <- poly_calc(0:5)
summary(p)
}
