% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{unique.polylist}
\alias{unique.polylist}
\title{Unique components}
\usage{
\method{unique}{polylist}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{A polylist object}

\item{incomparables}{Logical: as for the base function \code{unique}}

\item{...}{As for the base function \code{unique}}
}
\value{
A polylist object with no duplicated components
}
\description{
Remove duplicated polynomials in a polylist object
}
