% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{as.character.polynom}
\alias{as.character.polynom}
\title{Polynomial coercion to character}
\usage{
\method{as.character}{polynom}(x, variable = "x", decreasing = FALSE, ...)
}
\arguments{
\item{x}{The polynomial object in question}

\item{variable}{Character string: what variable name should be used?}

\item{decreasing}{Logical: in decreasing powers or increasing powers?}

\item{...}{Additional arguments (ignored as yet)}
}
\value{
A character string representation of the polynomial
}
\description{
Produce a text representation of a polynomial object
}
\examples{
p <- poly_from_zeros(-2:3)
as.character(p, "z", FALSE)
as.character(p, "z", TRUE)
parse(text = as.character(p, "z", TRUE))[[1]]
}
