% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomF.R
\name{as.function.polynom}
\alias{as.function.polynom}
\alias{as.function.polylist}
\title{Coercion to function}
\usage{
\method{as.function}{polynom}(x, variable = "x", ...)

\method{as.function}{polylist}(x, ...)
}
\arguments{
\item{x}{A polynomial object}

\item{variable}{Character string: what variable name should be used?}

\item{...}{Additional arguments}
}
\value{
An explicit R function evaluating the polynomial
}
\description{
PolynomF objects ARE functions, but this coercion method
creates from a polynomial object a pure function with the
coefficients fully exposed in the code and which evaluates
the polynomial more efficiently.
}
\examples{
p <- poly_from_zeros(-2:3)
p
as.function(p)
}
